move cert related functions of clab to cert.go
